Switch to profile access control
With the user base growing, it was no longer easy to provide access control on a per-user basis.

As such, a CustomerProfile model was implemented, it will store all authorities, and may be assigned to multiple users, thus giving all users assigned with it the same permissions.

By changing a profile permissions, all users assigned to it will have their permissions changed.

Some adjustments where also made to reduce the size of responses. The responses will now return the authorities only when needed.